add diungraph
