+++ pywhy_graphs/classes/__init__.py
@@ -1,6 +1,6 @@
 from . import timeseries
 from .admg import ADMG
-from .cpdag import CPDAG
+from .diungraph import CG, CPDAG
 from .intervention import IPAG, AugmentedGraph, PsiPAG
 from .pag import PAG
 from .timeseries import (
